+++ packages/undercut/src/iterable.js
@@ -1,6 +1,9 @@
 import { assert } from "./utils/assertions.js";
 import { isFunction } from "./utils/lang.js";
 
+/**
+ * A wrapper for generators to make them re-iterable.
+ */
 export default class Iterable {
 	constructor(createIterator) {
 		assert(isFunction(createIterator), "createIterator is required.");