+++ packages/undercut/src/pull/pull_core.test.js
@@ -1,8 +1,24 @@
-import { lineTargetOf } from "../utils/tests.js";
+import { lineTargetOf, targetOf, testOperation } from "../utils/tests.js";
 
+import { flatten } from "./operations/flatten.js";
 import { map } from "./operations/map.js";
+import { zip } from "./operations/zip.js";
 
-import { pull, createPullLine } from "./pull_core.js";
+import { pull, composeOperations, createPullLine } from "./pull_core.js";
+
+test("pull", () => {
+	expect(() => pull()).toThrow();
+	expect(() => pull(Array.from)).toThrow();
+	expect(() => pull(Array.from, [])).toThrow();
+	expect(() => pull(1, [], [])).toThrow();
+	expect(() => pull(Array.from, 2, [])).toThrow();
+	expect(() => pull(Array.from, [], 3)).toThrow();
+
+	expect(pull(Array.from, [], [])).toEqual([]);
+	expect(pull(Array.from, [], [6, 7])).toEqual([6, 7]);
+	expect(pull(Array.from, [map(x => x + 1)], [])).toEqual([]);
+	expect(pull(Array.from, [map(x => x * 0)], [3, 4])).toEqual([0, 0]);
+});
 
 test("createPullLine", () => {
 	expect(() => createPullLine()).toThrow();
@@ -21,16 +37,25 @@ test("createPullLine", () => {
 	expect([...pullLine, 7, ...pullLine]).toEqual([0, 2, 4, 7, 0, 2, 4]);
 });
 
-test("pull", () => {
-	expect(() => pull()).toThrow();
-	expect(() => pull(Array.from)).toThrow();
-	expect(() => pull(Array.from, [])).toThrow();
-	expect(() => pull(1, [], [])).toThrow();
-	expect(() => pull(Array.from, 2, [])).toThrow();
-	expect(() => pull(Array.from, [], 3)).toThrow();
+test("composeOperations", () => {
+	function interleave(...sources) {
+		const pipeline = [zip(...sources), flatten()];
 
-	expect(pull(Array.from, [], [])).toEqual([]);
-	expect(pull(Array.from, [], [6, 7])).toEqual([6, 7]);
-	expect(pull(Array.from, [map(x => x + 1)], [])).toEqual([]);
-	expect(pull(Array.from, [map(x => x * 0)], [3, 4])).toEqual([0, 0]);
+		return composeOperations(pipeline);
+	}
+
+	expect(targetOf(interleave([2, 4]), [1, 3])).toEqual([1, 2, 3, 4]);
+
+	const pipeline = [
+		testOperation(),
+		testOperation(),
+		interleave([2, 4]),
+		testOperation()
+	];
+
+	expect(pull(Array.from, pipeline, [1, 3])).toEqual([1, 2, 3, 4]);
+
+	const pullLine = createPullLine(pipeline, [1, 3]);
+
+	expect([...pullLine, 7, ...pullLine]).toEqual([1, 2, 3, 4, 7, 1, 2, 3, 4]);
 });